frontend/requirements.txt
pygame==2.5.2
python-socketio[client]==5.10.0
requests==2.31.0
firebase-admin==6.2.0
python-dotenv==1.0.0